refactor pulling scripts, not part of udal
