+++ contracts/pull_combined_logsheets.py
